% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_vs_normal_KL.R
\name{disease_vs_normal_KL}
\alias{disease_vs_normal_KL}
\title{Identify perturbed ligand-receptor interactions between two conditions}
\usage{
disease_vs_normal_KL(
  ligand_mat_disease,
  receptor_mat_disease,
  ligand_mat_normal,
  receptor_mat_normal,
  pseudo_count,
  method
)
}
\arguments{
\item{ligand_mat_disease}{ligand expression matrix in the disease/perturbed condition}

\item{receptor_mat_disease}{receptor expression matrix in the disease/perturbed condition. The same ligand-receptor pair should be in the same order as ligand_mat_disease}

\item{ligand_mat_normal}{ligand expression matrix in the normal/baseline condition. Same order as ligand_mat_disease}

\item{receptor_mat_normal}{receptor expression matrix in the normal/baseline coniditon. The same ligand-receptor pair should be in the same order as ligand_mat_normal}

\item{pseudo_count}{a small value added to the expression matrices to avoid dividing by 0.}

\item{method}{"product" or "min": ligand-receptor interaction score is either defined by multiplying ligand and receptor expression values or take the minimum}
}
\value{
KL divergence for the ligand-receptor interaction scores between disease vs. normal (or perturbed vs. baseline) conditions
}
\description{
This function calculates Kullback-Leibler divergence of ligand-receptor interaction scores between two conditions (e.g., disease vs. normal)
}
