% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_deg_net.R
\name{make_deg_net}
\alias{make_deg_net}
\title{Prioritize ligand-receptor interactions based on differential expression}
\usage{
make_deg_net(expression_data, expressed_net, fc_thresh, pseudo_count)
}
\arguments{
\item{expression_data}{expression matrix for ligand and receptor genes. 1st column "genes" (gene symbol, upper case); remaining columns: gene expression level}

\item{expressed_net}{the output from calling make_expressed_net.R function. To ensure the input network to both algorithms are the same.}

\item{fc_thresh}{fold change threshold above which a gene is considered enriched}

\item{pseudo_count}{a small value added to the expression matrices to avoid dividing by 0.}
}
\value{
ligand-receptor network where both ligand and receptor genes are uniquely enriched in one cell type (i.e, marker genes for each cell type)
}
\description{
A ligand-receptor pair is considerd when both the ligand and receptor genes are enriched in one cell type (e.g., expressed 3 fold higher in cell type A than each of the other cell types)
The resulting ligand-receptor network is then compared against talklr
}
