% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_expressed_net.R
\name{make_expressed_net}
\alias{make_expressed_net}
\title{Prioritize ligand-receptor interactions in a single condition}
\usage{
make_expressed_net(
  expression_data,
  expressed_thresh,
  receptor_ligand,
  KL_method,
  pseudo_count
)
}
\arguments{
\item{expression_data}{expression matrix for ligand and receptor genes. 1st column "genes" (gene symbol, upper case); remaining columns: gene expression level}

\item{expressed_thresh}{threshold above which a gene is considered expressed}

\item{receptor_ligand}{receptor-ligand interaction database from Ramilowski JA et al, Nat. Comm. 2015. Provided with the package}

\item{KL_method}{"product" or "min": ligand-receptor interaction score is either defined by multiplying ligand and receptor expression values or take the minimum}

\item{pseudo_count}{a small value added to the expression matrices to avoid dividing by 0.}
}
\value{
KL divergence for all expressed ligand-receptor pairs, as well as ligand and receptor expression values across cell types. You may want to sort from highest to lowest to examine the top ones.
}
\description{
Prioritize ligand-receptor interactions in a single condition
}
