% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_expressed_net_specify_expressed_genes.R
\name{make_expressed_net_specify_expressed_genes}
\alias{make_expressed_net_specify_expressed_genes}
\title{Prioritize ligand-receptor interactions in a single condition}
\usage{
make_expressed_net_specify_expressed_genes(
  expression_data,
  expressed_genes,
  receptor_ligand,
  KL_method
)
}
\arguments{
\item{expression_data}{expression matrix for ligand and receptor genes. 1st column "genes" (gene symbol, upper case); remaining columns: gene expression level}

\item{expressed_genes}{a character vector of expressed genes}

\item{receptor_ligand}{receptor-ligand interaction database from Ramilowski JA et al, Nat. Comm. 2015. Provided with the package}

\item{KL_method}{"product" or "min": ligand-receptor interaction score is either defined by multiplying ligand and receptor expression values or take the minimum}
}
\value{
KL divergence for all expressed ligand-receptor pairs, as well as ligand and receptor expression values across cell types. You may want to sort from highest to lowest to examine the top ones.
}
\description{
Instead of using an expression threshold, this function identifies interesting ligand-receptor pairs among a specific list of expressed genes provided by the user.
Run this function separately for perturbed and baseline conditions before running disease_vs_normal_KL.R, so that ligand and receptor matrices are in the same order.
}
