% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.interaction.KL.R
\name{pairwise.interaction.KL}
\alias{pairwise.interaction.KL}
\title{Calulate KL divergence between observed vs. expected interaction score distributions.}
\usage{
pairwise.interaction.KL(ligand_mat, receptor_mat, method)
}
\arguments{
\item{ligand_mat}{expression matrix for ligands}

\item{receptor_mat}{expression matrix for receptors. The same ligand-receptor pair should be in the same order as ligand_mat.}

\item{method}{"product" or "min": ligand-receptor interaction score is either defined by multiplying ligand and receptor expression values or take the minimum}
}
\value{
KL divergence for all expressed ligand-receptor pairs.
}
\description{
Expected interaction score distribution is that all cell types express the same level of ligand and receptor genes, thus the normalized probability is 1/n^2.
Observed distribution is the interaction scores based on actual expression values.
This is a helper function called by make_expressed_net.R or make_expressed_net_specify_expressed_genes.R
}
