% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GO_comparison.R
\name{plot_GO_comparison}
\alias{plot_GO_comparison}
\title{Make scatter plots of GO enrichment results}
\usage{
plot_GO_comparison(
  go_deg,
  go_product,
  interesting,
  fdr_thresh,
  diff_thresh,
  filename
)
}
\arguments{
\item{go_deg}{enrichment table for ligand-receptor pairs identified by the DEG approach. Two required columns are GO.ID and FDR}

\item{go_product}{enrichment table for ligand-receptor pairs identified by talklr}

\item{interesting}{a character vector of GO terms to be visualized}

\item{fdr_thresh}{FDR threshold above which a GO is considered significant}

\item{diff_thresh}{if differences of -log10(FDR) between the two appraoches are larger than this threshold, the GO terms will be labeled.}

\item{filename}{name of the file where the scatter plot will be saved.}
}
\value{
A scatter plot named filename
}
\description{
This function makes scatter plots of GO enrichment results. Diagonal line denotes equal enrichment.
GO terms with higher enrichment in talklr pairs are in green; GO terms with higher enrichment in DEG approach are in orange
Axis are drawn to scale. Longer axis means larger values.
}
