% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lr_wiring.R
\name{plot_lr_wiring}
\alias{plot_lr_wiring}
\title{Plot ligand-receptor wiring diagram}
\usage{
plot_lr_wiring(ligand_exprs, receptor_exprs, cell_labels)
}
\arguments{
\item{ligand_exprs}{ligand expression matrix}

\item{receptor_exprs}{receptor expression matrix}

\item{cell_labels}{character vector of cell type names}
}
\value{
a plot of the ligand-receptor wiring diagram
}
\description{
This function uses the iGraph package to plot the ligand-receptor wiring diagram among interacting cell types
}
