% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sig_bar.R
\name{plot_sig_bar}
\alias{plot_sig_bar}
\title{Make bar plots of ligand/receptor expression and interaction scors}
\usage{
plot_sig_bar(
  ligand_exprs,
  receptor_exprs,
  cell_labels,
  filenames,
  raw_color,
  norm_color,
  ligand_name,
  receptor_name,
  y_lim = 1
)
}
\arguments{
\item{ligand_exprs}{a numeric vector of ligand expression values across cell types}

\item{receptor_exprs}{a numeric vector of receptor expression values across cell types}

\item{cell_labels}{a character vector of cell type labels}

\item{filenames}{prefix for the name of the figures}

\item{raw_color}{color for the bar plot of raw interaction scores}

\item{norm_color}{color for the bar plot of normalized interaction scores}

\item{ligand_name}{name of the ligand}

\item{receptor_name}{name of the receptor}

\item{ylim}{maximum y axis for the normalized interaction scores.Default is 1}
}
\value{
returns bar plots of ligand and receptor expression, as well as raw and normalized interaction scores
}
\description{
This plot makes bar plots for ligand expression, receptor expression, raw interaction scores and normalized interaction scores for a specific ligand-receptor pair
}
\examples{
plot_sig_bar(c(5,1,200),c(5,1,200),c("A","B","C"),"Figure4A_baseline",'slateblue3','slateblue1',"EGF","EGFR",1)
}
